+++ burn-wgpu/src/codegen/mod.rs
@@ -1,0 +1,13 @@
+mod body;
+mod function;
+mod kernel;
+mod operator;
+mod shader;
+mod variable;
+
+pub(crate) use body::*;
+pub(crate) use function::*;
+pub(crate) use kernel::*;
+pub(crate) use operator::*;
+pub(crate) use shader::*;
+pub(crate) use variable::*;
