--- burn-wgpu/src/fusion/codegen/mod.rs
@@ -1,11 +1,0 @@
-mod body;
-mod function;
-mod operator;
-mod shader;
-mod variable;
-
-pub use body::*;
-pub use function::*;
-pub use operator::*;
-pub use shader::*;
-pub use variable::*;
