+++ burn-wgpu/src/fusion/mod.rs
@@ -2,7 +2,6 @@ mod base;
 mod elemwise;
 
 pub(crate) mod cache;
-pub(crate) mod codegen;
 pub(crate) mod kernel;
 
 pub use base::*;
